FlightSevlet refactoring & Start working on frontEnd (Index , Flight, Flights)
