+++ .idea/dataSources.local.xml
@@ -7,35 +7,17 @@
       </database-info>
       <case-sensitivity plain-identifiers="lower" quoted-identifiers="exact" />
       <secret-storage>master_key</secret-storage>
-      <user-name>imads</user-name>
       <schema-mapping>
         <introspection-scope>
-          <node negative="1">
-            <node kind="database" qname="@">
-              <node kind="schema" qname="@" />
-            </node>
-            <node kind="database">
-              <name qname="postgres" />
-              <name qname="AirAfrika" />
-              <node kind="schema" negative="1" />
-            </node>
+          <node kind="database" qname="postgres">
+            <node kind="schema" negative="1" />
           </node>
         </introspection-scope>
       </schema-mapping>
     </data-source>
     <data-source name="hibernate.cfg.xml/Hibernate" uuid="013d4e9b-9e3b-41cb-9bba-915b42cb5bca">
-      <database-info product="PostgreSQL" version="15.4 (Homebrew)" jdbc-version="4.2" driver-name="PostgreSQL JDBC Driver" driver-version="42.6.0" dbms="POSTGRES" exact-version="15.4" exact-driver-version="42.6">
-        <identifier-quote-string>&quot;</identifier-quote-string>
-      </database-info>
-      <case-sensitivity plain-identifiers="lower" quoted-identifiers="exact" />
-      <user-name>imads</user-name>
-      <schema-mapping>
-        <introspection-scope>
-          <node kind="database" qname="@">
-            <node kind="schema" qname="@" />
-          </node>
-        </introspection-scope>
-      </schema-mapping>
+      <database-info product="" version="" jdbc-version="" driver-name="" driver-version="" dbms="POSTGRES" exact-version="0" />
+      <schema-mapping />
     </data-source>
   </component>
 </project>
