+++ .idea/dataSources/f954423f-831a-4d92-9088-bf8a5e893d51/storage_v2/_src_/database/AirAfrika.PMH_Qw/schema/pg_catalog.0S1ZNQ.meta
@@ -1,0 +1,2 @@
+#n:pg_catalog
+!<md> [null, 0, null, null, -2147483648, -2147483648]
